+++ crates/miden-lib/src/transaction/kernel_procedures.rs
@@ -48,9 +48,9 @@ pub const KERNEL_PROCEDURES: [Word; 51] = [
     // account_remove_asset
     word!("0x7b965e458a962667a9cdc54a6677fd0c5a573bd7b265ccc33775a8a985aeead5"),
     // account_get_balance
-    word!("0x52233827fc91ef50a5dc09f74d3176011d244fd80e291ae751b64bdaa2c6cf75"),
+    word!("0x6f9c831083e90227f161ca5a23bd475e9fac14a3f7a0b4d5b8a49e47301973a4"),
     // account_get_initial_balance
-    word!("0xf0db3924f3e2d677a51924b09ecef8a12416a6ceb09fadd39785bb4f685cab66"),
+    word!("0x31d3842969889dc731790fe89d9053a7d97a125d05a057712012a19aa668943b"),
     // account_has_non_fungible_asset
     word!("0xf975c799cffebf8565a8479475fe04c1832ef2a7484c4a2a42bbf7d8a340d649"),
     // account_compute_delta_commitment